Move code and code injector up
