+++ java/src/main/kotlin/io/spine/validation/java/JavaValidationRenderer.kt
@@ -37,8 +37,6 @@ import io.spine.protodata.java.render.findClass
 import io.spine.protodata.render.SourceFile
 import io.spine.protodata.render.SourceFileSet
 import io.spine.tools.code.Java
-import io.spine.validation.java.generate.MessageValidationCode
-import io.spine.validation.java.generate.ValidationCodeInjector
 import io.spine.validation.java.generate.option.DistinctGenerator
 import io.spine.validation.java.generate.option.GoesGenerator
 import io.spine.validation.java.generate.option.PatternGenerator
